Delete (get/post) completed.
The Delete for both 'get' and 'post' methods inside the Superhero Controller has been completed. Dev tested and stepped through the method to ensure that removal of a superhero was successful, which it was. Dev will continue to do further testing of these completed methods but will start working on the Details method which will fetch and display the details of a specific superhero.